kvserver/closedts: pull test helpers into sender_helper_test.go
This commit extracts test helpers into a separate file, sender_helper_test.go.
While it doesn't alter existing behavior, this change prepares for future
commits that will introduce additional helper utilities and tests which make
this file too large.

Epic: none
Release note: none
